refactor(organization): extract EmployeeTreeNode and buildTree functions for better modularity and readability
