+++ app/users/dashboard/organizations/[id]/page.tsx
@@ -5,8 +5,6 @@ import { useState, use } from 'react';
 import { Button } from '@/components/ui/button';
 import { Badge } from '@/components/ui/badge';
 import { Department, getDepartmentLabel } from '@/types/employee/departments';
-import type { Employee } from '@/types/employee';
-import { EmployeeAvatar } from '@/features/employee/components/employee-avatar';
 import {
   Card,
   CardContent,
@@ -39,71 +37,8 @@ import { format } from 'date-fns';
 import { useOrganization as useOrganizationQuery } from '@/hooks/organization/use-organizations';
 import { useOrganization } from '@/components/providers/organization-provider';
 import { toast } from '@/lib/styles/toast-styles';
-import { LeavePoliciesManager } from '@/features/leave/components/leave-policies-manager';
-
-// ---------------------------------------------------------------------------
-// Hierarchy tree helpers
-// ---------------------------------------------------------------------------
-
-type TreeNode = Employee & { children: TreeNode[] };
-
-function buildTree(employees: Employee[]): TreeNode[] {
-  const nodeMap = new Map<number, TreeNode>();
-  for (const emp of employees) {
-    if (emp.id !== undefined) nodeMap.set(emp.id, { ...emp, children: [] });
-  }
-  const roots: TreeNode[] = [];
-  for (const node of nodeMap.values()) {
-    if (node.managerId !== undefined && nodeMap.has(node.managerId)) {
-      nodeMap.get(node.managerId)!.children.push(node);
-    } else {
-      roots.push(node);
-    }
-  }
-  return roots;
-}
-
-function EmployeeTreeNode({
-  node,
-  isLast,
-}: {
-  node: TreeNode;
-  isLast: boolean;
-}) {
-  return (
-    <div className="relative">
-      <div className="flex items-start gap-3 rounded-lg p-2 transition-colors hover:bg-zinc-50 dark:hover:bg-zinc-800/50">
-        <EmployeeAvatar employee={node} size="sm" />
-        <div className="min-w-0 flex-1">
-          <p className="truncate text-sm font-medium text-zinc-900 dark:text-zinc-100">
-            {node.name}
-          </p>
-          <p className="truncate text-xs text-zinc-500 dark:text-zinc-400">
-            {node.designation}
-          </p>
-        </div>
-        {node.children.length > 0 && (
-          <Badge variant="outline" className="shrink-0 text-xs">
-            {node.children.length}{' '}
-            {node.children.length === 1 ? 'report' : 'reports'}
-          </Badge>
-        )}
-      </div>
-
-      {node.children.length > 0 && (
-        <div className="mt-1 ml-5 border-l-2 border-zinc-200 pl-4 dark:border-zinc-700">
-          {node.children.map((child, i) => (
-            <EmployeeTreeNode
-              key={child.id}
-              node={child}
-              isLast={i === node.children.length - 1}
-            />
-          ))}
-        </div>
-      )}
-    </div>
-  );
-}
+import { buildTree } from '@/lib/utils/hierarchy';
+import { EmployeeTreeNode } from '@/features/employee/components/employee-tree-node';
 
 interface OrganizationDetailPageProps {
   params: Promise<{
@@ -632,12 +567,8 @@ export default function OrganizationDetailPage({
 
                     {treeRoots.length > 0 ? (
                       <div className="space-y-3">
-                        {treeRoots.map((root, i) => (
-                          <EmployeeTreeNode
-                            key={root.id}
-                            node={root}
-                            isLast={i === treeRoots.length - 1}
-                          />
+                        {treeRoots.map((root) => (
+                          <EmployeeTreeNode key={root.id} node={root} />
                         ))}
                       </div>
                     ) : (
