+++ features/employee/index.ts
@@ -4,3 +4,4 @@ export { EmployeeProjectsCell } from './components/employee-projects-cell';
 export { EmployeeStatusBadge } from './components/employee-status-badge';
 export { EmployeeAvatar } from './components/employee-avatar';
 export { EmployeeTable } from './components/employee-table';
+export { EmployeeTreeNode } from './components/employee-tree-node';
